requirements.txt
flask
mongoengine
tldextract
psycopg2-binary
fastapi
uvicorn